feat: add notifications

+++ src/main/java/ru/rsreu/exchangeofthings/constant/GlobalOptions.java
@@ -4,5 +4,5 @@ package ru.rsreu.exchangeofthings.constant;
  * Global options of project.
  */
 public class GlobalOptions {
-    public static final long SESSION_TIME_LIVE = 1000 * 60 * 60;
+    public static final int SESSION_TIME_LIVE = 1000 * 60 * 60;
 }
